feat: Update adonisrc.ts to import commands and providers
The code changes in this commit update the `adonisrc.ts` file to import additional commands and providers. This enhances the functionality of the AdonisJS application by including commands from the `@adonisjs/core`, `@adonisjs/lucid`, and `@adonisjs/bouncer` packages. It also adds providers for CORS, database, authentication, and bouncer. These changes improve the overall capabilities of the application.

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>