feat: Add migration for creating company_professionals table
This commit adds a migration file for creating the `company_professionals` table in the database. The table has columns for `company_id` and `professional_id`, both of which are foreign keys referencing the `companies` and `professionals` tables respectively. The migration also includes a unique constraint on the combination of `company_id` and `professional_id`. This migration improves the database structure by establishing the relationship between companies and professionals.

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>