feat: Update logger configuration to enable/disable based on environment
This commit updates the logger configuration in the `logger.ts` file. The `enabled` property of the `app` logger is now set based on the value of the `NODE_ENV` environment variable. If the `NODE_ENV` is not equal to 'test', the logger will be enabled. This change ensures that the logger is only active in non-test environments, improving performance and reducing unnecessary logging in tests.

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>