refactor: Update import statement for Profession model in professional_seeder.ts
This commit updates the import statement for the Profession model in the professional_seeder.ts file. The previous import was using the deprecated ProfessionalType model, which has been replaced with the Profession model. This change ensures that the correct model is used for creating a new profession in the seeder.

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>